moved repository, azureimage, and getsubscriptions to the correct places within deleteImage
