+++ utils/azureCredentialsManager.ts
@@ -3,11 +3,6 @@ import { Registry } from 'azure-arm-containerregistry/lib/models';
 import { ResourceManagementClient, SubscriptionClient, SubscriptionModels } from 'azure-arm-resource';
 import { ResourceGroup, ResourceGroupListResult } from "azure-arm-resource/lib/resource/models";
 import { ServiceClientCredentials } from 'ms-rest';
-import request = require('request-promise');
-import * as vscode from "vscode";
-import { AzureImageNode, AzureRegistryNode, AzureRepositoryNode } from '../explorer/models/azureRegistryNodes';
-import * as azureUtils from '../explorer/utils/azureUtils';
-import { AzureImage, getSub, Repository } from '../explorer/utils/azureUtils';
 import * as ContainerModels from '../node_modules/azure-arm-containerregistry/lib/models';
 import { AzureAccount, AzureSession } from '../typings/azure-account.api';
 import { AsyncPool } from '../utils/asyncpool';
